+++ lib/microstrain_mag_cal/calibration.cpp
@@ -10,7 +10,7 @@
 namespace microstrain_mag_cal
 {
     /// @brief Returns a fit result that leaves the calibration unchanged (doesn't apply).
-    FitResult FitResult::noCorrection(const Error &error)
+    FitResult FitResult::noCorrection(const Error error)
     {
         return {Eigen::Matrix3d::Identity(), Eigen::Vector3d::Zero(), error};
     }
@@ -146,10 +146,13 @@ namespace microstrain_mag_cal
         solver.parameters.maxfev = MAX_ITERATIONS;
         solver.parameters.xtol = TOLERANCE;
 
-        // Optimize
+        // Optimize.
+        // Status codes outside of this range indicate that the solver was not able to converge.
+        // Status codes within this range indicate that the solver converged early (i.e. no more
+        // error reduction can be made).
         if (const Eigen::LevenbergMarquardtSpace::Status status = solver.minimize(parameters);
-            status != Eigen::LevenbergMarquardtSpace::Status::RelativeErrorTooSmall &&
-            status != Eigen::LevenbergMarquardtSpace::Status::RelativeReductionTooSmall)
+            status <= Eigen::LevenbergMarquardtSpace::Status::ImproperInputParameters ||
+            status >= Eigen::LevenbergMarquardtSpace::Status::TooManyFunctionEvaluation)
         {
             return FitResult::Error::FIT_OPTIMIZATION_DID_NOT_CONVERGE;
         }